+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 ### Changed
 - When installed, the plugin now saves `Default.json`, `Full.json` and `Simple.json` TinyMCE config files in a Craft install's `config/tinymce` directory, if the files didn't already exist
 
+### Fixed
+- Fixed a bug where the `craftcms/html-field` package dependency wasn't set
+
 ## 1.0.1 - 2022-09-26
 
 ### Fixed